Use the same logic to get original message
