+++ src/events/request/RequestEventHandler.ts
@@ -107,9 +107,8 @@ export default class RequestEventHandler implements EventHandler<'message'> {
 				.setColor( 'BLUE' )
 				.setAuthor( origin.author.tag, origin.author.avatarURL() )
 				.setDescription( this.replaceTicketReferencesWithRichLinks( origin.content, regex ) )
+				.addField( 'Note', 'WIP', true )
 				.addField( 'Go To', `[Message](${ origin.url }) in ${ origin.channel }`, true )
-				.addField( 'Channel', origin.channel.id, true )
-				.addField( 'Message', origin.id, true )
 				.setTimestamp( new Date() );
 
 			const response = BotConfig.request.prependResponseMessage == PrependResponseMessageType.Always
